job #12345 Fix conditionality on R35
R35 between Graphical Element and Layer is unconditional on the "Layer"
side (i.e. every graphical element must exist in at least one layer).
BridgePoint doesn't enforce this. By default elements are not in any
layer (which is treated as a default always visible layer). It has been
changed to "conditional" on the "Layer" side to match the behavior
BridgePoint is already following.