job #12560 fix content asssist in 'elif' blocks
After the introduction of the 'then' keyword, content assist was broken
for the first statement within 'elif' blocks. This also revealed that
content assist was not added when the 'then' and 'loop' keywords are
used. This change fixes the failing tests and adds support for content
assist when using the new keywords.