Update the actual TOC for reference material
It seems that bpalref.book.htm is obsolete, so updating it (as I did
during my first attempt) was rather pointless.  Opening a browser on
that file allowed me to convince myself that I had everything in place,
but this was just an illusion.
Testing with a locally produced build showed that all the files for the
new Relationships section were in place, but poking the link in the Help
facility produced a blank page because I had failed to notice the _real_
TOC for reference material resides in topics_Reference.xml.  This commit
updates that file, and I have tested the change by copying the updated
file into my installation and navigating to the updated Relationships
section from within the Help facility.
Separately, I opened https://support.onefact.net/issues/12563 suggesting
the deletion of a couple obsolete files, but that work is outside the
scope of this change and not included in this commit.