In work - will tag for research
Saving this version which still allows re-assign of associative class in
editor; but which does not render correctly.
When an associative is re-assigned in the editor, the original link
remains and a new link, with no destination, is drawn.
Saving for subsequent possible exploration of this graphics misbehavior.
A further commit will disable re-assignment - and, thus, inconvenience
further research.