job #12073 - Subsequent Imports Fails For Welcome
This fix removes the code that was checking for a shell text that is
removed after the first call to ProjectUtilities::importExistingProject.

It is assumed that a change to the Eclipse Wizard handling modified the
behavior described in a comment in
ProjectUtilities::importModelUsingWizard associated with
dialog.getShell().dispose(). The comment states, "If we do not
explicitly dispose the shell here it hangs-around and the shell name
is simply "import" which conflicts with other types of imports causing
the wrong shell to be selected (dts0100806930)."