Remove tablular association editor
The tabular association editor was an experiment from ~2017 that was
never adopted by anyone.  It relies on code generated by the wizard
voodoo from PEI data specified in core/sql/context_menu_pei.sql.  When
the PEIs for the formalize and unformalize CMEs are removed, the tabular
association editor code fails to compile.