12676 - Clear ReferencePaths
When the textual canvas editor was created there were listeners
to keep the ReferencePath and ObjectReferences in sync with
loaded/unloaded elements.  The current approach is more to load
from scratch every time.

To support this we add a utility method that can remove ReferencePath
and ObjectReferences that are a suffix to the parent.  This is now
called on load of any textual graphics file.